requirements-no-torch.txt
# Install mlx-whisper separately with --no-deps.
# Upstream declares torch, but Muesli's runtime path uses MLX directly.

mlx==0.31.0
mlx-metal==0.31.0
mlx-audio==0.4.0
numpy==2.4.2
numba==0.64.0
scipy==1.17.1
tiktoken==0.12.0
more-itertools==10.8.0
huggingface_hub==1.5.0
httpx==0.28.1
tqdm==4.67.3

sounddevice==0.5.5
rumps==0.4.0
pynput==1.8.1
pyperclip==1.11.0

pyobjc-framework-ApplicationServices==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-CoreMedia==12.1
pyobjc-framework-EventKit==12.1
pyobjc-framework-Quartz==12.1
pyobjc-framework-ScreenCaptureKit==12.1
